core/env.py
import os

from dotenv import load_dotenv

load_dotenv()

OPENROUTER_API_KEY = os.environ.get("OPENROUTER_API_KEY")
if OPENROUTER_API_KEY is None:
    raise ValueError(f"Environment variable 'OPENROUTER_API_KEY' was not found in the .env file.")
